Rename to stitchmd
"mdreduce" isn't specific enough.
"stitchmd" at least conveys that we'll be stitching Markdown files
together.
