+++ internal/goldast/walk.go
@@ -2,7 +2,7 @@ package goldast
 
 import (
 	"github.com/yuin/goldmark/ast"
-	"go.abhg.dev/mdreduce/internal/pos"
+	"go.abhg.dev/stitchmd/internal/pos"
 )
 
 // Visitor visits individual nodes in a Goldmark AST.
